ml/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch
torchvision
pillow
numpy>=2.0.0
python-dotenv
gunicorn
opencv-python
pdf2image
piexif
